+++ stan/math/prim/arr/meta/VectorBuilderHelper.hpp
@@ -1,0 +1,25 @@
+#ifndef STAN_MATH_PRIM_ARR_META_VECTORBUILDER_HELPER_HPP
+#define STAN_MATH_PRIM_ARR_META_VECTORBUILDER_HELPER_HPP
+
+#include <stan/math/prim/scal/meta/VectorBuilderHelper.hpp>
+#include <stdexcept>
+#include <vector>
+
+namespace stan {
+
+  /**
+   * Template specialization for using a vector
+   */
+  template<typename T1>
+  class VectorBuilderHelper<T1, true, true> {
+  private:
+    std::vector<T1> x_;
+  public:
+    explicit VectorBuilderHelper(size_t n) : x_(n) { }
+    T1& operator[](size_t i) {
+      return x_[i];
+    }
+  };
+
+}
+#endif
